@@ -12,11 +12,13 @@
 //
 //===----------------------------------------------------------------------===//
 
-import XCTest
+import Foundation
+import Testing
 
 @testable import AWSLambdaEvents
 
-class CloudwatchTests: XCTestCase {
+@Suite
+struct CloudwatchTests {
     static func eventBody(type: String, details: String) -> String {
         """
         {
@@ -34,75 +36,56 @@ class CloudwatchTests: XCTestCase {
         """
     }
 
-    func testScheduledEventFromJSON() {
+    @Test func scheduledEventFromJSON() throws {
         let eventBody = CloudwatchTests.eventBody(type: CloudwatchDetails.Scheduled.name, details: "{}")
         let data = eventBody.data(using: .utf8)!
-        var maybeEvent: CloudwatchScheduledEvent?
-        XCTAssertNoThrow(maybeEvent = try JSONDecoder().decode(CloudwatchScheduledEvent.self, from: data))
+        let event = try JSONDecoder().decode(CloudwatchScheduledEvent.self, from: data)
 
-        guard let event = maybeEvent else {
-            return XCTFail("Expected to have an event")
-        }
-
-        XCTAssertEqual(event.id, "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
-        XCTAssertEqual(event.source, "aws.events")
-        XCTAssertEqual(event.accountId, "123456789012")
-        XCTAssertEqual(event.time, Date(timeIntervalSince1970: 0))
-        XCTAssertEqual(event.region, .us_east_1)
-        XCTAssertEqual(event.resources, ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
+        #expect(event.id == "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
+        #expect(event.source == "aws.events")
+        #expect(event.accountId == "123456789012")
+        #expect(event.time == Date(timeIntervalSince1970: 0))
+        #expect(event.region == .us_east_1)
+        #expect(event.resources == ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
     }
 
-    func testEC2InstanceStateChangeNotificationEventFromJSON() {
+    @Test func ec2InstanceStateChangeNotificationEventFromJSON() throws {
         let eventBody = CloudwatchTests.eventBody(
             type: CloudwatchDetails.EC2.InstanceStateChangeNotification.name,
             details: "{ \"instance-id\": \"0\", \"state\": \"stopping\" }"
         )
         let data = eventBody.data(using: .utf8)!
-        var maybeEvent: CloudwatchEC2InstanceStateChangeNotificationEvent?
-        XCTAssertNoThrow(
-            maybeEvent = try JSONDecoder().decode(CloudwatchEC2InstanceStateChangeNotificationEvent.self, from: data)
-        )
+        let event = try JSONDecoder().decode(CloudwatchEC2InstanceStateChangeNotificationEvent.self, from: data)
 
-        guard let event = maybeEvent else {
-            return XCTFail("Expected to have an event")
-        }
-
-        XCTAssertEqual(event.id, "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
-        XCTAssertEqual(event.source, "aws.events")
-        XCTAssertEqual(event.accountId, "123456789012")
-        XCTAssertEqual(event.time, Date(timeIntervalSince1970: 0))
-        XCTAssertEqual(event.region, .us_east_1)
-        XCTAssertEqual(event.resources, ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
-        XCTAssertEqual(event.detail.instanceId, "0")
-        XCTAssertEqual(event.detail.state, .stopping)
+        #expect(event.id == "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
+        #expect(event.source == "aws.events")
+        #expect(event.accountId == "123456789012")
+        #expect(event.time == Date(timeIntervalSince1970: 0))
+        #expect(event.region == .us_east_1)
+        #expect(event.resources == ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
+        #expect(event.detail.instanceId == "0")
+        #expect(event.detail.state == .stopping)
     }
 
-    func testEC2SpotInstanceInterruptionNoticeEventFromJSON() {
+    @Test func ec2SpotInstanceInterruptionNoticeEventFromJSON() throws {
         let eventBody = CloudwatchTests.eventBody(
             type: CloudwatchDetails.EC2.SpotInstanceInterruptionNotice.name,
             details: "{ \"instance-id\": \"0\", \"instance-action\": \"terminate\" }"
         )
         let data = eventBody.data(using: .utf8)!
-        var maybeEvent: CloudwatchEC2SpotInstanceInterruptionNoticeEvent?
-        XCTAssertNoThrow(
-            maybeEvent = try JSONDecoder().decode(CloudwatchEC2SpotInstanceInterruptionNoticeEvent.self, from: data)
-        )
+        let event = try JSONDecoder().decode(CloudwatchEC2SpotInstanceInterruptionNoticeEvent.self, from: data)
 
-        guard let event = maybeEvent else {
-            return XCTFail("Expected to have an event")
-        }
-
-        XCTAssertEqual(event.id, "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
-        XCTAssertEqual(event.source, "aws.events")
-        XCTAssertEqual(event.accountId, "123456789012")
-        XCTAssertEqual(event.time, Date(timeIntervalSince1970: 0))
-        XCTAssertEqual(event.region, .us_east_1)
-        XCTAssertEqual(event.resources, ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
-        XCTAssertEqual(event.detail.instanceId, "0")
-        XCTAssertEqual(event.detail.action, .terminate)
+        #expect(event.id == "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
+        #expect(event.source == "aws.events")
+        #expect(event.accountId == "123456789012")
+        #expect(event.time == Date(timeIntervalSince1970: 0))
+        #expect(event.region == .us_east_1)
+        #expect(event.resources == ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
+        #expect(event.detail.instanceId == "0")
+        #expect(event.detail.action == .terminate)
     }
 
-    func testS3ObjectCreatedEventFromJSON() {
+    @Test func s3ObjectCreatedEventFromJSON() throws {
         let eventBody = CloudwatchTests.eventBody(
             type: CloudwatchDetails.S3.ObjectCreatedNotification.name,
             details:
@@ -110,35 +93,28 @@ class CloudwatchTests: XCTestCase {
 
         )
         let data = eventBody.data(using: .utf8)!
-        var maybeEvent: CloudWatchS3ObjectCreatedNotificationEvent?
-        XCTAssertNoThrow(
-            maybeEvent = try JSONDecoder().decode(CloudWatchS3ObjectCreatedNotificationEvent.self, from: data)
-        )
+        let event = try JSONDecoder().decode(CloudWatchS3ObjectCreatedNotificationEvent.self, from: data)
 
-        guard let event = maybeEvent else {
-            return XCTFail("Expected to have an event")
-        }
-
-        XCTAssertEqual(event.id, "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
-        XCTAssertEqual(event.source, "aws.events")
-        XCTAssertEqual(event.accountId, "123456789012")
-        XCTAssertEqual(event.time, Date(timeIntervalSince1970: 0))
-        XCTAssertEqual(event.region, .us_east_1)
-        XCTAssertEqual(event.resources, ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
-        XCTAssertEqual(event.detail.version, "0")
-        XCTAssertEqual(event.detail.bucket.name, "amzn-s3-demo-bucket1")
-        XCTAssertEqual(event.detail.object.key, "example-key")
-        XCTAssertEqual(event.detail.object.size, 5)
-        XCTAssertEqual(event.detail.object.etag, "b1946ac92492d2347c6235b4d2611184")
-        XCTAssertEqual(event.detail.object.versionId, "IYV3p45BT0ac8hjHg1houSdS1a.Mro8e")
-        XCTAssertEqual(event.detail.object.sequencer, "617f08299329d189")
-        XCTAssertEqual(event.detail.requestId, "N4N7GDK58NMKJ12R")
-        XCTAssertEqual(event.detail.requester, "123456789012")
-        XCTAssertEqual(event.detail.sourceIpAddress, "1.2.3.4")
-        XCTAssertEqual(event.detail.reason, .putObject)
+        #expect(event.id == "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
+        #expect(event.source == "aws.events")
+        #expect(event.accountId == "123456789012")
+        #expect(event.time == Date(timeIntervalSince1970: 0))
+        #expect(event.region == .us_east_1)
+        #expect(event.resources == ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
+        #expect(event.detail.version == "0")
+        #expect(event.detail.bucket.name == "amzn-s3-demo-bucket1")
+        #expect(event.detail.object.key == "example-key")
+        #expect(event.detail.object.size == 5)
+        #expect(event.detail.object.etag == "b1946ac92492d2347c6235b4d2611184")
+        #expect(event.detail.object.versionId == "IYV3p45BT0ac8hjHg1houSdS1a.Mro8e")
+        #expect(event.detail.object.sequencer == "617f08299329d189")
+        #expect(event.detail.requestId == "N4N7GDK58NMKJ12R")
+        #expect(event.detail.requester == "123456789012")
+        #expect(event.detail.sourceIpAddress == "1.2.3.4")
+        #expect(event.detail.reason == .putObject)
     }
 
-    func testS3ObjectDeletedEventFromJSON() {
+    @Test func s3ObjectDeletedEventFromJSON() throws {
         let eventBody = CloudwatchTests.eventBody(
             type: CloudwatchDetails.S3.ObjectDeletedNotification.name,
             details:
@@ -146,35 +122,28 @@ class CloudwatchTests: XCTestCase {
 
         )
         let data = eventBody.data(using: .utf8)!
-        var maybeEvent: CloudWatchS3ObjectDeletedNotificationEvent?
-        XCTAssertNoThrow(
-            maybeEvent = try JSONDecoder().decode(CloudWatchS3ObjectDeletedNotificationEvent.self, from: data)
-        )
+        let event = try JSONDecoder().decode(CloudWatchS3ObjectDeletedNotificationEvent.self, from: data)
 
-        guard let event = maybeEvent else {
-            return XCTFail("Expected to have an event")
-        }
-
-        XCTAssertEqual(event.id, "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
-        XCTAssertEqual(event.source, "aws.events")
-        XCTAssertEqual(event.accountId, "123456789012")
-        XCTAssertEqual(event.time, Date(timeIntervalSince1970: 0))
-        XCTAssertEqual(event.region, .us_east_1)
-        XCTAssertEqual(event.resources, ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
-        XCTAssertEqual(event.detail.version, "0")
-        XCTAssertEqual(event.detail.bucket.name, "amzn-s3-demo-bucket1")
-        XCTAssertEqual(event.detail.object.key, "example-key")
-        XCTAssertEqual(event.detail.object.etag, "d41d8cd98f00b204e9800998ecf8427e")
-        XCTAssertEqual(event.detail.object.versionId, "1QW9g1Z99LUNbvaaYVpW9xDlOLU.qxgF")
-        XCTAssertEqual(event.detail.object.sequencer, "617f0837b476e463")
-        XCTAssertEqual(event.detail.requestId, "0BH729840619AG5K")
-        XCTAssertEqual(event.detail.requester, "123456789012")
-        XCTAssertEqual(event.detail.sourceIpAddress, "1.2.3.4")
-        XCTAssertEqual(event.detail.reason, .deleteObject)
-        XCTAssertEqual(event.detail.deletionType, .deleteMarkerCreated)
+        #expect(event.id == "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
+        #expect(event.source == "aws.events")
+        #expect(event.accountId == "123456789012")
+        #expect(event.time == Date(timeIntervalSince1970: 0))
+        #expect(event.region == .us_east_1)
+        #expect(event.resources == ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
+        #expect(event.detail.version == "0")
+        #expect(event.detail.bucket.name == "amzn-s3-demo-bucket1")
+        #expect(event.detail.object.key == "example-key")
+        #expect(event.detail.object.etag == "d41d8cd98f00b204e9800998ecf8427e")
+        #expect(event.detail.object.versionId == "1QW9g1Z99LUNbvaaYVpW9xDlOLU.qxgF")
+        #expect(event.detail.object.sequencer == "617f0837b476e463")
+        #expect(event.detail.requestId == "0BH729840619AG5K")
+        #expect(event.detail.requester == "123456789012")
+        #expect(event.detail.sourceIpAddress == "1.2.3.4")
+        #expect(event.detail.reason == .deleteObject)
+        #expect(event.detail.deletionType == .deleteMarkerCreated)
     }
 
-    func testS3ObjectRestoreCompletedEventFromJSON() {
+    @Test func s3ObjectRestoreCompletedEventFromJSON() throws {
         let eventBody = CloudwatchTests.eventBody(
             type: CloudwatchDetails.S3.ObjectRestoreCompletedNotification.name,
             details:
@@ -182,34 +151,27 @@ class CloudwatchTests: XCTestCase {
 
         )
         let data = eventBody.data(using: .utf8)!
-        var maybeEvent: CloudWatchS3ObjectRestoreCompletedNotificationEvent?
-        XCTAssertNoThrow(
-            maybeEvent = try JSONDecoder().decode(CloudWatchS3ObjectRestoreCompletedNotificationEvent.self, from: data)
-        )
+        let event = try JSONDecoder().decode(CloudWatchS3ObjectRestoreCompletedNotificationEvent.self, from: data)
 
-        guard let event = maybeEvent else {
-            return XCTFail("Expected to have an event")
-        }
-
-        XCTAssertEqual(event.id, "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
-        XCTAssertEqual(event.source, "aws.events")
-        XCTAssertEqual(event.accountId, "123456789012")
-        XCTAssertEqual(event.time, Date(timeIntervalSince1970: 0))
-        XCTAssertEqual(event.region, .us_east_1)
-        XCTAssertEqual(event.resources, ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
-        XCTAssertEqual(event.detail.version, "0")
-        XCTAssertEqual(event.detail.bucket.name, "amzn-s3-demo-bucket1")
-        XCTAssertEqual(event.detail.object.key, "example-key")
-        XCTAssertEqual(event.detail.object.size, 5)
-        XCTAssertEqual(event.detail.object.etag, "b1946ac92492d2347c6235b4d2611184")
-        XCTAssertEqual(event.detail.object.versionId, "KKsjUC1.6gIjqtvhfg5AdMI0eCePIiT3")
-        XCTAssertEqual(event.detail.requestId, "189F19CB7FB1B6A4")
-        XCTAssertEqual(event.detail.requester, "s3.amazonaws.com")
-        XCTAssertEqual(event.detail.restoreExpiryTime.description, "2021-11-13 00:00:00 +0000")
-        XCTAssertEqual(event.detail.sourceStorageClass, .glacier)
+        #expect(event.id == "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
+        #expect(event.source == "aws.events")
+        #expect(event.accountId == "123456789012")
+        #expect(event.time == Date(timeIntervalSince1970: 0))
+        #expect(event.region == .us_east_1)
+        #expect(event.resources == ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
+        #expect(event.detail.version == "0")
+        #expect(event.detail.bucket.name == "amzn-s3-demo-bucket1")
+        #expect(event.detail.object.key == "example-key")
+        #expect(event.detail.object.size == 5)
+        #expect(event.detail.object.etag == "b1946ac92492d2347c6235b4d2611184")
+        #expect(event.detail.object.versionId == "KKsjUC1.6gIjqtvhfg5AdMI0eCePIiT3")
+        #expect(event.detail.requestId == "189F19CB7FB1B6A4")
+        #expect(event.detail.requester == "s3.amazonaws.com")
+        #expect(event.detail.restoreExpiryTime.description == "2021-11-13 00:00:00 +0000")
+        #expect(event.detail.sourceStorageClass == .glacier)
     }
 
-    func testCustomEventFromJSON() {
+    @Test func customEventFromJSON() throws {
         struct Custom: CloudwatchDetail {
             public static let name = "Custom"
 
@@ -218,38 +180,33 @@ class CloudwatchTests: XCTestCase {
 
         let eventBody = CloudwatchTests.eventBody(type: Custom.name, details: "{ \"name\": \"foo\" }")
         let data = eventBody.data(using: .utf8)!
-        var maybeEvent: CloudwatchEvent<Custom>?
-        XCTAssertNoThrow(maybeEvent = try JSONDecoder().decode(CloudwatchEvent<Custom>.self, from: data))
+        let event = try JSONDecoder().decode(CloudwatchEvent<Custom>.self, from: data)
 
-        guard let event = maybeEvent else {
-            return XCTFail("Expected to have an event")
-        }
-
-        XCTAssertEqual(event.id, "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
-        XCTAssertEqual(event.source, "aws.events")
-        XCTAssertEqual(event.accountId, "123456789012")
-        XCTAssertEqual(event.time, Date(timeIntervalSince1970: 0))
-        XCTAssertEqual(event.region, .us_east_1)
-        XCTAssertEqual(event.resources, ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
-        XCTAssertEqual(event.detail.name, "foo")
+        #expect(event.id == "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c")
+        #expect(event.source == "aws.events")
+        #expect(event.accountId == "123456789012")
+        #expect(event.time == Date(timeIntervalSince1970: 0))
+        #expect(event.region == .us_east_1)
+        #expect(event.resources == ["arn:aws:events:us-east-1:123456789012:rule/ExampleRule"])
+        #expect(event.detail.name == "foo")
     }
 
-    func testUnregistredType() {
+    @Test func unregistredType() throws {
         let eventBody = CloudwatchTests.eventBody(type: UUID().uuidString, details: "{}")
         let data = eventBody.data(using: .utf8)!
-        XCTAssertThrowsError(try JSONDecoder().decode(CloudwatchScheduledEvent.self, from: data)) { error in
-            XCTAssert(error is CloudwatchDetails.TypeMismatch, "expected DetailTypeMismatch but received \(error)")
+        #expect(throws: CloudwatchDetails.TypeMismatch.self) {
+            try JSONDecoder().decode(CloudwatchScheduledEvent.self, from: data)
         }
     }
 
-    func testTypeMismatch() {
+    @Test func typeMismatch() throws {
         let eventBody = CloudwatchTests.eventBody(
             type: CloudwatchDetails.EC2.InstanceStateChangeNotification.name,
             details: "{ \"instance-id\": \"0\", \"state\": \"stopping\" }"
         )
         let data = eventBody.data(using: .utf8)!
-        XCTAssertThrowsError(try JSONDecoder().decode(CloudwatchScheduledEvent.self, from: data)) { error in
-            XCTAssert(error is CloudwatchDetails.TypeMismatch, "expected DetailTypeMismatch but received \(error)")
+        #expect(throws: CloudwatchDetails.TypeMismatch.self) {
+            try JSONDecoder().decode(CloudwatchScheduledEvent.self, from: data)
         }
     }
 }
